ref(bot): move merge.message tests into their own files
This will help with a future change I'm making that touches some of these tests.
